requirements.txt
python-telegram-bot==20.0
telethon
python-dotenv
